+++ python/aistore/sdk/const.py
@@ -79,6 +79,7 @@ QPARAM_NEW_CUSTOM = "set-new-custom"
 # etl
 QPARAM_ETL_NAME = "etl_name"
 QPARAM_ETL_ARGS = "etl_args"
+QPARAM_ETL_PIPELINE = "etl_pipeline"
 # etl websocket
 ETL_WS_PIPELINE = "pipeline"
 ETL_WS_FQN = "fqn"