+++ python/aistore/sdk/etl/etl_config.py
@@ -4,7 +4,8 @@
 from dataclasses import dataclass
 from typing import Dict, Optional, Union
 from json import dumps as json_dumps
-from aistore.sdk.const import QPARAM_ETL_NAME, QPARAM_ETL_ARGS
+from aistore.sdk.const import QPARAM_ETL_NAME, QPARAM_ETL_ARGS, QPARAM_ETL_PIPELINE
+from aistore.sdk.etl.etl import Etl
 
 
 @dataclass
@@ -13,12 +14,11 @@ class ETLConfig:
     Configuration for ETL operations.
 
     Attributes:
-        name (str): Name of the ETL pipeline.
+        name (Optional[Union[str, Etl]]): Name of the ETL pipeline or an Etl object representing a pipeline.
         args (Optional[Union[str, Dict]]): Optional parameters for configuring the ETL pipeline.
             Can be provided as:
                 - A string for simple arguments.
                 - A dictionary for structured key-value configurations.
-            Defaults to an empty dictionary.
 
     Example:
         etl_config = ETLConfig(
@@ -27,7 +27,7 @@ class ETLConfig:
         )
     """
 
-    name: str
+    name: Optional[Union[str, Etl]] = None
     args: Optional[Union[str, Dict]] = None
 
     def update_qparams(self, params: Dict[str, str]) -> None:
@@ -37,7 +37,15 @@ class ETLConfig:
         Args:
             params (Dict[str, str]): Query parameters dict to modify
         """
-        params[QPARAM_ETL_NAME] = self.name
+        if not self.name:
+            return
+
+        if isinstance(self.name, Etl):
+            params[QPARAM_ETL_NAME] = self.name.name
+            params[QPARAM_ETL_PIPELINE] = ",".join(self.name.pipeline)
+        else:
+            params[QPARAM_ETL_NAME] = self.name
+
         if self.args:
             etl_args = (
                 json_dumps(self.args, separators=(",", ":"))